Add manifest file
This commit adds a file which lists all files which are considered part
of "conmath" and are therefore subject to the LPPL v1.3. All other files
are simply utility files for creating the package and do not need to be
licensed with the LPPL v1.3.

Using a manifest file avoids having mismatching information about the
files of the work as it is documented in one place only.