impl: Conditionally load amssymb package
For the future addition of a d'Alembert operator, a box/square symbol is
needed.
Load the amssymb package if the option "no-vector" has not been given,
so that the \square symbol is available.

This commit also adds a section to the documentation about the
dependencies of the package and updates the README to mention amssymb as
dependency.