impl: Add document level command centering math
Sometimes a document author also needs to explicitly center something
along the math axis.

This commit provides the command \VCenterMath, which passes its
arguments to \conmath_vcenter:n to achieve this.
An error is raised if the command is invoked outside of math mode,
to avoid cryptic error message by TeX.