impl: Add vcenter option
The package should be able to vertically center certain math operators,
where the automatic (vertical) centering of \mathop fails.

As I do not know (yet) if this is the right thing to do, from a
typographical point of view, this shall be controlled by a package
option.

This commit adds the option vcenter, to achieve this.