Replace conmath with moremath
After talking with people from CTAN and TeXLive it was recommended to
rename the package.

Karl Berry came up with the great name "moremath" that is going to be
name of this package in the future.

This commit renames the package from "conmath" to "moremath",
replacing every occurrence of the package name with moremath
and also replacing the prefix "conmath" with moremath.